# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation de l'API pour la gestion des articles",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Liste des articles",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "responses": {
                    "200": {
                        "description": "Liste des articles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Créer un article",
                "operationId": "e328bed23caf889ad982ad3f1180bccb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mon premier article"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Ceci est une description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé avec succès"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Afficher un article spécifique",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'article"
                    },
                    "404": {
                        "description": "Article non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Mettre à jour un article",
                "operationId": "2717d5a8e3b1f16637aded325347f9ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Titre mis à jour"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description mise à jour"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article mis à jour avec succès"
                    },
                    "422": {
                        "description": "Validation échouée"
                    },
                    "404": {
                        "description": "Article non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Supprimer un article",
                "operationId": "97f34592611a1dd53a086865e29706c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Article supprimé avec succès"
                    },
                    "404": {
                        "description": "Article non trouvé"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        }
    ]
}